Pin functionality update 3.8
basic pin functionality is working. I might expand on it later and make it so that it will say to user please try again if they make a mistake. For now im going to continue onto the transaction history section as this is the final part. Then after brush up on the software and make it more user friendly as well as trouble shoot